requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
annotated-types==0.5.0
    # via pydantic
anyio[trio]==3.7.1
    # via
    #   -r requirements/dev.in
    #   httpcore
    #   starlette
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
async-generator==1.10
    # via trio
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via
    #   outcome
    #   trio
backcall==0.2.0
    # via ipython
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
click==8.1.5
    # via uvicorn
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
ecdsa==0.18.0
    # via python-jose
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
executing==1.2.0
    # via stack-data
fastapi==0.100.0
    # via -r requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements/dev.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   trio
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r requirements/dev.in
ipython==8.14.0
    # via
    #   -r requirements/dev.in
    #   ipdb
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
outcome==1.2.0
    # via trio
packaging==23.1
    # via pytest
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycryptodome==3.18.0
    # via python-jose
pydantic==2.0.3
    # via
    #   -r requirements/requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.3.0
    # via pydantic
pydantic-settings==2.0.2
    # via -r requirements/requirements.in
pygments==2.15.1
    # via ipython
pytest==7.4.0
    # via
    #   -r requirements/dev.in
    #   pytest-asyncio
pytest-asyncio==0.21.1
    # via -r requirements/dev.in
python-dotenv==1.0.0
    # via pydantic-settings
python-jose[pycryptodome]==3.3.0
    # via -r requirements/requirements.in
redis==4.6.0
    # via -r requirements/requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   asttokens
    #   ecdsa
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
stack-data==0.6.2
    # via ipython
starlette==0.27.0
    # via
    #   -r requirements/requirements.in
    #   fastapi
tomli==2.0.1
    # via
    #   ipdb
    #   pytest
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
trio==0.21.0
    # via anyio
typing-extensions==4.7.1
    # via
    #   fastapi
    #   ipython
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
uvicorn==0.23.1
    # via -r requirements/requirements.in
wcwidth==0.2.6
    # via prompt-toolkit
